src/requirements.txt
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gitdb==4.0.10
GitPython==3.1.36
greenlet==2.0.1
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.7
python-dotenv==0.21.0
requests==2.28.2
smmap==5.0.0
SQLAlchemy==1.4.46
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.2
WTForms==3.0.1
